requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
ortools>=9.8 # Specify minimum version if known
numpy
sqlalchemy # Assuming this is still used somewhere

# GPU Acceleration (Optional - Install matching your CUDA version)
# If you have a compatible NVIDIA GPU and CUDA toolkit installed,
cupy-cuda118
# Other examples: cupy-cuda116, cupy-cuda117, cupy-cuda12x

# Numba (Used for CPU fallback and GPU kernels)
numba>=0.57 # Specify minimum version if known

# Optional for CPU detection/concurrency
psutil
